Type Definition rust_ssvm::types::ffi::evmc_emit_log_fn

type evmc_emit_log_fn = Option<unsafe extern "C" fn(*mut evmc_context, *const evmc_address, *const u8, usize, *const evmc_bytes32, usize)>;

Log callback function.

This callback function is used by an EVM to inform about a LOG that happened during an EVM bytecode execution. @param context The pointer to the Host execution context. @see ::evmc_context. @param address The address of the contract that generated the log. @param data The pointer to unindexed data attached to the log. @param data_size The length of the data. @param topics The pointer to the array of topics attached to the log. @param topics_count The number of the topics. Valid values are between 0 and 4 inclusively.