Type Definition rust_ssvm::types::ffi::evmc_execute_fn

type evmc_execute_fn = Option<unsafe extern "C" fn(*mut evmc_instance, *mut evmc_context, evmc_revision, *const evmc_message, *const u8, usize) -> evmc_result>;

Executes the given code using the input from the message.

This function MAY be invoked multiple times for a single VM instance.

@param instance The VM instance. This argument MUST NOT be NULL. @param context The pointer to the Host execution context to be passed to the Host interface methods (::evmc_host_interface). This argument MUST NOT be NULL unless the @p instance has the ::EVMC_CAPABILITY_PRECOMPILES capability. @param rev The requested EVM specification revision. @param msg The call parameters. See ::evmc_message. This argument MUST NOT be NULL. @param code The reference to the code to be executed. This argument MAY be NULL. @param code_size The length of the code. If @p code is NULL this argument MUST be 0. @return The execution result.