Type Definition rust_ssvm::types::ffi::evmc_get_block_hash_fn

pub type evmc_get_block_hash_fn = Option<unsafe extern "C" fn(*mut evmc_context, i64) -> evmc_bytes32>;

Get block hash callback function.

This callback function is used by a VM to query the hash of the header of the given block. If the information about the requested block is not available, then this is signalled by returning null bytes.

@param context The pointer to the Host execution context. @param number The block number. @return The block hash or null bytes if the information about the block is not available.