Type Definition rust_ssvm::types::ffi::evmc_set_tracer_fn

pub type evmc_set_tracer_fn = Option<unsafe extern "C" fn(*mut evmc_instance, Option<unsafe extern "C" fn(*mut evmc_tracer_context, usize, evmc_status_code, i64, usize, *const evmc_bytes32, usize, usize, usize, *const u8)>, *mut evmc_tracer_context)>;

Sets the EVM instruction tracer.

When the tracer is set in the EVM instance, the EVM SHOULD call back the tracer with information about instructions execution in the EVM. @see ::evmc_trace_callback.

This will overwrite the previous settings (the callback and the context).

@deprecated Deprecated since EVMC 6.3, see evmc_instance::set_tracer().

@param instance The EVM instance. @param callback The tracer callback function. This argument MAY be NULL to disable previously set tracer. @param context The Client-side tracer context. This argument MAY be NULL in case the tracer does not require any context. This argument MUST be NULL if the callback argument is NULL.